+++ PyAPIReference/extra.py
@@ -2,6 +2,8 @@ import os
 from importlib.util import spec_from_file_location, module_from_spec
 from PyQt5.QtWidgets import QAction, QDialog, QLabel, QVBoxLayout
 
+TAB = "&nbsp;" * 4 
+
 def create_qaction(menu, text: str, shortcut: str="", callback: callable=lambda: print("No callback"), parent=None) -> QAction:
 	"""This function will create a QAction and return it"""
 	action = QAction(parent) # Create a qaction in the window (self)
@@ -25,9 +27,13 @@ def get_module_from_path(path: str):
 
 	return module
 
-def convert_to_code_block(string: str) -> str:
-	result = "<pre style='background-color: #404040; color: white;'>"
-	result += str(string)
-	result += "</pre>"
+def convert_to_code_block(string: str, stylesheet: str="background-color: #484848; color: white;") -> str:
+	if not isinstance(string, str):
+		return convert_to_code_block("None", stylesheet=stylesheet)
+
+
+	result = f"<span style='{stylesheet}'>"
+	result += str(string).replace("\t", TAB).replace("\n", "<br>")
+	result += "</span>"
 
 	return result